+++ package.json
@@ -52,6 +52,7 @@
     "heroku-postbuild": "npm run build && (echo \"api.tasks.tasks['stats:generate'].run(api, {}, (e) => {console.log('DONE', e)})\";sleep 10) | npm run ah console"
   },
   "dependencies": {
+    "@sentry/node": "^5.9.0",
     "actionhero": "^20.0.0",
     "archiver": "^3.1.1",
     "bcrypt": "^3.0.7",
@@ -89,7 +90,6 @@
     "pg": "^7.14.0",
     "pg-hstore": "^2.3.3",
     "querystring": "^0.2.0",
-    "raven": "^2.6.4",
     "sequelize": "^5.21.2",
     "sequelize-cli": "^5.5.1",
     "sharp": "^0.23.3",
