+++ server/config/sentry.js
@@ -1,0 +1,7 @@
+exports.default = {
+  sentry: function (api) {
+    return {
+      dsn: process.env.SENTRY_DSN
+    }
+  }
+}
